+++ packages/demo/src/modules/users/pages/UserForm.vue
@@ -2,15 +2,10 @@
 /**
  * UserForm - Create/Edit user page
  *
- * Uses useFormPageBuilder with generateFields() for auto-resolved reference options.
+ * Uses useFormPageBuilder with FormInput for auto-rendered fields.
  */
+import { useFormPageBuilder, FormPage, FormField, FormInput } from 'qdadm'
 
-import { useFormPageBuilder, FormPage, FormField } from 'qdadm'
-import InputText from 'primevue/inputtext'
-import Password from 'primevue/password'
-import Select from 'primevue/select'
-
-// ============ FORM BUILDER ============
 const form = useFormPageBuilder({ entity: 'users' })
 form.generateFields()
 form.addSaveAction({ andClose: true })
@@ -22,22 +17,7 @@ form.addDeleteAction()
     <template #fields>
       <div class="form-grid">
         <FormField v-for="f in form.fields.value" :key="f.name" :name="f.name" :label="f.label">
-          <InputText v-if="f.type === 'text'" v-model="form.data.value[f.name]" class="w-full" />
-          <Password
-            v-else-if="f.type === 'password'"
-            v-model="form.data.value[f.name]"
-            class="w-full"
-            :feedback="false"
-            toggleMask
-          />
-          <Select
-            v-else-if="f.type === 'select'"
-            v-model="form.data.value[f.name]"
-            :options="f.options"
-            optionLabel="label"
-            optionValue="value"
-            class="w-full"
-          />
+          <FormInput :field="f" v-model="form.data.value[f.name]" />
         </FormField>
       </div>
     </template>